Create seprate service for validatePasswordRestToken
